+++ src/main/java/org/hvdw/jexiftoolgui/facades/SystemPropertyFacade.java
@@ -3,10 +3,10 @@ package org.hvdw.jexiftoolgui.facades;
 public class SystemPropertyFacade {
     public enum SystemPropertyKey {
         LINE_SEPARATOR("line.separator"),
-        FILE_SEPARATOR(" file.separator"),
+        FILE_SEPARATOR("file.separator"),
         OS_NAME("os.name"),
         OS_ARCH("os.arch"),
-        OS_VERSION(" os.version"),
+        OS_VERSION("os.version"),
         USER_DIR("user.dir"),
         USER_HOME("user.home"),
         USER_NAME("user.name"),